feat(Examples): Wrap H-Keyring as a Master Key
This allows the H-Keyring to be used with a Caching CMM,
but only if the Algorithm suite does not support Digital Signatures,
only if there is one Encrypted Data Key,
and only if a static Branch Key ID is used.

Though, it would not be hard to add multiple
Encrypted Data Key support,
the other caveats are harder to work around.