fix: remove pre-mature Deprecation notices
While the Master Key Provider and Master Key interfaces do not have a long term future,
they are still supported, and safe to use.

They will not be removed in version 3.x.
They MAY be removed in a future Major Version,
but only when AWS is preparing to release that new Major Version should AWS publish these deprecation notices.